+++ src/components/group-study/review/group/detail/group-review-card.tsx
@@ -4,7 +4,7 @@ import { cn } from '@/components/common/ui/(shadcn)/lib/utils';
 import Avatar from '@/components/common/ui/avatar';
 import type { GroupStudyExperienceReviewListItem } from '@/hooks/queries/group-study/group-study-review-api';
 import { formatDateTimeDot } from '@/utils/time';
-import { useExpandableContent } from '../../../_components/use-expandable-content';
+import { useExpandableContent } from '@/components/group-study/review/use-expandable-content';
 
 export default function GroupReviewCard({
   review,
