+++ src/components/group-study/review/group/detail/satisfaction-section.tsx
