+++ src/components/group-study/review/group/member-study-card.tsx
