+++ src/components/group-study/review/use-expandable-content.ts
